requirements.txt
anthropic>=0.40.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
